Most relevant functions

InPhyNet.netnj — Function

netnj(D::Matrix{Float64}, constraints::Vector{HybridNetwork}, namelist::AbstractVector{<:AbstractString})

Creates a supernetwork based on the constraint trees/networks in constraints and distance matrix D.

Arguments

D: distance matrix relating pairs of taxa. This can be generated from estimated gene trees with calculateAGID

InPhyNet.calculateAGID — Function

calculateAGID(Ns::AbstractVector{HybridNetwork})

Calculates the Average Gene tree Internode Distance for all pairs of taxa across all networks in Ns.

Returns

AGID matrix

names of taxa corresponding to rows/columns in the AGID matrix

All documented functions

InPhyNet.TIEWARNING

InPhyNet.Graph

InPhyNet.addrootedge!

InPhyNet.calculateAGIC

InPhyNet.calculateAGID

InPhyNet.calculateAGID

InPhyNet.check_constraint

InPhyNet.check_constraints

InPhyNet.check_ptable

InPhyNet.check_quartets

InPhyNet.connectnodes!

InPhyNet.decomposeFromQuartets

InPhyNet.decomposeFromQuartets

InPhyNet.expandhybsubsetsD!

InPhyNet.find_not_in_path_edge_vector

InPhyNet.find_valid_node_path

InPhyNet.findsiblingpairs

InPhyNet.findvalidpairs

InPhyNet.fuseredundantedges!

InPhyNet.gather_hyb_descendants_outside_of_path

InPhyNet.getLeavesUnderEdge

InPhyNet.getLongestBranch

InPhyNet.getMidSplitEdge

InPhyNet.getsiblingcandidates

InPhyNet.inany

InPhyNet.incr_pick_sequence!

InPhyNet.inphynet_deep_search

InPhyNet.internodecount

InPhyNet.internodedistance

InPhyNet.kconnected

InPhyNet.log_edge_path_retics_from_node

InPhyNet.major_mrca

InPhyNet.majorinternodecount

InPhyNet.majorinternodedistance

InPhyNet.mergesubnets!

InPhyNet.minimumuniquesubset

InPhyNet.netnj

InPhyNet.netnj

InPhyNet.parsequartets

InPhyNet.parsequartets

InPhyNet.placeretics!

InPhyNet.pruneTruthFromDecomp

InPhyNet.removeredundantedges!

InPhyNet.requiredhybridsubsets

InPhyNet.root_constraints!

InPhyNet.sateIIdecomp

InPhyNet.sateIdecomp

InPhyNet.splitAtEdge

InPhyNet.updateconstraints!

InPhyNet.vecreplace!

InPhyNet.TIEWARNING — Constant

findoptQ(D::Matrix{Float64}, idxpairs::Vector{Tuple{<:Integer, <:Integer}})

Finds the minimizer (i, j) among all pairs (i, j) in idxpairs for Q, a matrix computed from D.

InPhyNet.Graph — Method

Converts the tree/network net into a SimpleGraph to leverage already implemented pathfinding algorithms.

Arguments

includeminoredges (default=true): if true, the entire network is translated to a graph. Otherwise, only tree-like edges (other than those in alwaysinclude) are retained.

alwaysinclude (default=nothing): edges that should always be included in the graph, regardless of the value of includeminoredges

withweights (default=false): return a set of weights corresponding to branch lengths as well

removeredundantedgecost (default=false): if true, finds all the nodes that would be removed via removeredundantedges! and sets their weights in W to 0

InPhyNet.addrootedge! — Method

Helper function that adds edges coming out from the root in net back into graph.

InPhyNet.calculateAGIC — Method

calculateAGIC(Ns::AbstractVector{HybridNetwork})

Calculates the Average Gene tree Internode Counts for all pairs of taxa across all networks in Ns.

Returns

AGIC matrix

names of taxa corresponding to rows/columns in the AGID matrix

InPhyNet.calculateAGID — Method

calculateAGID(Ns::AbstractVector{HybridNetwork})

Calculates the Average Gene tree Internode Distance for all pairs of taxa across all networks in Ns.

Returns

AGID matrix

names of taxa corresponding to rows/columns in the AGID matrix

InPhyNet.check_constraint — Method

Checks validity of a single input constraint networks. Checks include:

All nodes have exactly 3 edges except the root (unless the network is a single taxa)

Reticulations do not lead directly into other reticulations

InPhyNet.check_constraints — Method

Checks validity of input constraints. So far, the only check is to make sure that all nodes have exactly 3 edges except for the root.

InPhyNet.check_ptable — Method

Checks the input table to make sure that it is Holmes-Bonferroni adjusted. TODO: if it isn't, manually adjust it instead of throwing an error.

InPhyNet.check_quartets — Method

Makes sure the vector of abstract strings (that we call quartets) each have 4 unique entries.

InPhyNet.connectnodes! — Method

Deals with all the overhead of connecting two nodes with an edge. Returns the edge used to connect the nodes.

InPhyNet.decomposeFromQuartets — Method

Helper function that feeds into decomposeFromQuartets(namelist::ListOfNames, R::QuartetVector)

InPhyNet.decomposeFromQuartets — Method

Decomposes the set of all taxa into disjoint subsets intelligently based on a roughly estimated species tree and significant quartets from MSCquartets.

Arguments

T: the quickly estimated, rough species tree

R: the set of statistically significant quartets, as obtained by MSCquartets

Return Value

Returns a tuple. The first entry is list of subsets that are required in order to indentify all relevant reticulations. The second entry is the remaining taxa that aren't in the first entry.

InPhyNet.expandhybsubsetsD! — Method

expandhybsubsetsD!(hybsubs, other, quartets, D; maxlen::Int64=9)

Helper function; hybsubs is the subset of taxa selected for hybridization identification, other is all other taxa which are currently unplaced. This function expands the taxa in other appropriately into various subsets in hybsubs based on distance matrix D.

InPhyNet.find_not_in_path_edge_vector — Method

Helper function for when an edge (notinpath_edge) appears on but not within a merge path. Finds and returns: 1. Whether the edge should be logged with direction "to" or "from" 2. Whether the edge should be associated with subnetedgei or subnetedgej

InPhyNet.find_valid_node_path — Method

Helper function that finds a path in net from nodei to nodej. If that path has 2 or more reticulations in it, the cost to include a hybrid in the path is incrementally increased until only 1 reticulation is in the path.

InPhyNet.findsiblingpairs — Method

findsiblingpairs(net::HybridNetwork)

Finds sibling pairings for all the leaves in a single network. These pairs are valid for net but may not be valid when the other constraint networks are also considered. Returns a vector of tuples of nodes corresponding to siblings.

InPhyNet.findvalidpairs — Method

findvalidpairs(constraints::Vector{HybridNetwork}, namelist::AbstractVector{<:AbstractString})

Finds all valid sibling pairs among the constraint networks.

InPhyNet.fuseredundantedges! — Method

Helper function to remove redundant edges (e.g. A –> (internal node) –> (internal node)) that arise from one case of the function mergeconstraintnodes!.

Redundant nodes in this case will always have only two edges while not being the root.

InPhyNet.gather_hyb_descendants_outside_of_path — Method

Helper function used at one point in mergeconstraintnodes!. Gets all of the descendants of node that do not have any children and are not in path

InPhyNet.getLeavesUnderEdge — Method

Gets all the leaf nodes underneath edge edge.

InPhyNet.getLongestBranch — Method

Gets the longest edge in tre.

InPhyNet.getMidSplitEdge — Method

Finds a branch in tre that, when pruned on, splits the tree roughly in half.

InPhyNet.getsiblingcandidates — Method

A helper function that finds other nodes in the net that leaf belongs to that may be its sibling. This function sees past reticulations, so that e.g. in ((A,(B,#H1)),((C)#H1,D)) B and C will be marked as siblings.

InPhyNet.inany — Method

Helper function that finds whether name is listed in any subset in subsets

InPhyNet.incr_pick_sequence! — Method

Increments the merging sequence, i.e. [1, 1, 2, 1, 2, 3] –> [1, 1, 1, 2, 1, 2, 4]. Or, if the last idx had only 3 options, [1, 1, 2, 1, 2, 3] –> [1, 1, 2, 1, 3]. Notice in the second example length(pick_sequence) went down by 1.

InPhyNet.inphynet_deep_search — Method

Deep search for netnj

InPhyNet.internodecount — Method

internodecount(N::HybridNetwork; namelist::Union{Nothing,<:AbstractVector{String}}=nothing)

Calculates internode counts between all pairs of taxa in network N.

InPhyNet.internodedistance — Method

internodedistance(N::HybridNetwork; namelist::Union{Nothing,<:AbstractVector{String}}=nothing)

Calculates internode distance between all pairs of taxa in network N.

InPhyNet.kconnected — Method

Takes minsubset, a subset defined in the function minimumuniquesubset, and, for each entry i of minsubset, gathers the k-1 taxa that i appears with most often across all quartets.

InPhyNet.log_edge_path_retics_from_node — Method

Helper function

InPhyNet.major_mrca — Method

Gets the MRCA of nodei and nodej along their major tree.

InPhyNet.majorinternodecount — Method

majorinternodecount(N::HybridNetwork)

Calculates major internode counts between all pairs of taxa in the major displayed tree of network N. If N is a tree, then internode counts are as expected for a tree.

InPhyNet.majorinternodedistance — Method

majorinternodedistance(N::HybridNetwork)

Calculates internode distances between all pairs of taxa in the major displayed tree of network N. If N is a tree, then internode distances are as expected for a tree.

InPhyNet.mergesubnets! — Method

Merges two unrooted SubNets into a single unrooted SubNet Returns the merged subnet AND the edges in the merged net corresponding to where reticulations will be placed if there are corresponding reticulations discovered in the constraints.

IMPORTANT: this function should NOT make copies of edges or nodes, object references are stored and used in the main merging algorithm

InPhyNet.minimumuniquesubset — Method

Finds the smallest subset S of names in namelist such that every quartet in R has at least one of its taxa in S.

InPhyNet.netnj — Method

netnj(D::Matrix{Float64}, constraints::Vector{HybridNetwork}, namelist::AbstractVector{<:AbstractString})

Creates a supernetwork based on the constraint trees/networks in constraints and distance matrix D.

Arguments

D: distance matrix relating pairs of taxa. This can be generated from estimated gene trees with calculateAGID

InPhyNet.parsequartets — Method

Reads output from MSCquartets in R and parses it so that it can be used efficiently with our julia code.

InPhyNet.parsequartets — Method

Reads output from MSCquartets in R and parses it so that it can be used efficiently with our julia code.

InPhyNet.placeretics! — Method

Updates net to include the reticulations that we've kept track of along the way in our algo but haven't placed yet.

InPhyNet.pruneTruthFromDecomp — Method

Returns a list of networks where each entry is a copy of truenet pruned to include only the taxa named in each respective set in subsets.

Example

mynetwork = readTopology("((t1,t2),(t3,t4));")
pruneTruthFromDecomp(mynetwork, [["t1", "t2"], ["t3", "t4"]])
> 2-element Vector{HybridNetwork}: (t1, t2); and (t3, t4);

InPhyNet.removeredundantedges! — Method

Helper function that removes redundant edges in the graph that exist when some hybrids are pruned from the graph.

InPhyNet.requiredhybridsubsets — Method

Current base subset decomposition schema. First, we find the smallest subset S of names such that all quartets have at least one taxa in the subset. Then, for each entry of S, we gather the set of 3 other taxa it appears with most often in the quartets in R.

The idea behind these subsets is to guarantee identifiability of hybridizations, hence why they are required.

InPhyNet.root_constraints! — Method

If any constraints are unrooted, picks a root node from any node directly under the root (somewhat randomly).

InPhyNet.sateIIdecomp — Method

Performs subset decomposition as outlined in SATe-II on inittree.

InPhyNet.sateIdecomp — Method

Performs subset decomposition as outlines in SATe-I on inittree. Returns vector of vector of names, e.g. [["a", "b"], ["c", "d", "e"]]

InPhyNet.splitAtEdge — Method

Splits tre into two trees at its longest branch.

InPhyNet.updateconstraints! — Method

Updates constraint networks after (i, j) with names (nodenamei, nodenamej) have been merged. Also updates reticmap to keep track of any reticulations that get removed in this process.

By convention we keep nodenamei and replace node names with nodenamej

InPhyNet.vecreplace! — Method

Helper function for manipulating graph adjacency lists.