Most relevant functions

NetMerge.netnj — Function

netnj(D::Matrix{Float64}, constraints::Vector{HybridNetwork})

Creates a super-network based on the constraint trees/networks in constraints and distance matrix D.

Arguments

D: distance matrix relating pairs of taxa. This can be generated from estimated gene trees with calculateAGID

NetMerge.calculateAGID — Function

calculateAGID(Ns::AbstractVector{HybridNetwork})

Calculates the Average Gene tree Internode Distance for all pairs of taxa across all networks in Ns.

Returns

AGID matrix

names of taxa corresponding to rows/columns in the AGID matrix

All documented functions

NetMerge.TIEWARNING

NetMerge.Graph

NetMerge.addrootedge!

NetMerge.calculateAGID

NetMerge.calculateAGID

NetMerge.check_constraints

NetMerge.check_ptable

NetMerge.check_quartets

NetMerge.connectnodes!

NetMerge.decomposeFromQuartets

NetMerge.decomposeFromQuartets

NetMerge.expandhybsubsetsD!

NetMerge.findsiblingpairs

NetMerge.findvalidpairs

NetMerge.fuseredundantedges!

NetMerge.getsiblingcandidates

NetMerge.inany

NetMerge.internodedistance

NetMerge.kconnected

NetMerge.majorinternodedistance

NetMerge.mergesubnets!

NetMerge.minimumuniquesubset

NetMerge.netnj

NetMerge.netnj

NetMerge.parsequartets

NetMerge.parsequartets

NetMerge.placeretics!

NetMerge.removeredundantedges!

NetMerge.requiredhybridsubsets

NetMerge.updateconstraints!

NetMerge.vecreplace!

NetMerge.TIEWARNING — Constant

findoptQ(D::Matrix{Float64}, idxpairs::Vector{Tuple{<:Integer, <:Integer}})

Finds the minimizer (i, j) among all pairs (i, j) in idxpairs for Q, a matrix computed from D.

NetMerge.Graph — Method

Converts the tree/network net into a SimpleGraph to leverage already implemented pathfinding algorithms.

Arguments

includeminoredges (default=true): if true, the entire network is translated to a graph. Otherwise, only tree-like edges (other than those in alwaysinclude) are retained.

alwaysinclude (default=nothing): edges that should always be included in the graph, regardless of the value of includeminoredges

NetMerge.addrootedge! — Method

Helper function that adds edges coming out from the root in net back into graph.

NetMerge.calculateAGID — Method

calculateAGID(Ns::AbstractVector{HybridNetwork})

Calculates the Average Gene tree Internode Distance for all pairs of taxa across all networks in Ns.

Returns

AGID matrix

names of taxa corresponding to rows/columns in the AGID matrix

NetMerge.check_constraints — Method

Checks validity of input constraints. So far, the only check is to make sure that all nodes have exactly 3 edges except for the root.

NetMerge.check_ptable — Method

Checks the input table to make sure that it is Holmes-Bonferroni adjusted. TODO: if it isn't, manually adjust it instead of throwing an error.

NetMerge.check_quartets — Method

Makes sure the vector of abstract strings (that we call quartets) each have 4 unique entries.

NetMerge.connectnodes! — Method

Deals with all the overhead of connecting two nodes with an edge. Returns the edge used to connect the nodes.

NetMerge.decomposeFromQuartets — Method

Helper function that feeds into decomposeFromQuartets(namelist::ListOfNames, R::QuartetVector)

NetMerge.decomposeFromQuartets — Method

Decomposes the set of all taxa into disjoint subsets intelligently based on a roughly estimated species tree and significant quartets from MSCquartets.

Arguments

T: the quickly estimated, rough species tree

R: the set of statistically significant quartets, as obtained by MSCquartets

Return Value

Returns a tuple. The first entry is list of subsets that are required in order to indentify all relevant reticulations. The second entry is the remaining taxa that aren't in the first entry.

NetMerge.expandhybsubsetsD! — Method

expandhybsubsetsD!(hybsubs, other, quartets, D; maxlen::Int64=9)

Helper function; hybsubs is the subset of taxa selected for hybridization identification, other is all other taxa which are currently unplaced. This function expands the taxa in other appropriately into various subsets in hybsubs based on distance matrix D.

NetMerge.findsiblingpairs — Method

findsiblingpairs(net::HybridNetwork)

Finds sibling pairings for all the leaves in a single network. These pairs are valid for net but may not be valid when the other constraint networks are also considered. Returns a vector of tuples of nodes corresponding to siblings.

NetMerge.findvalidpairs — Method

findvalidpairs(constraints::Vector{HybridNetwork}, namelist::AbstractVector{<:AbstractString})

Finds all valid sibling pairs among the constraint networks.

NetMerge.fuseredundantedges! — Method

Helper function to remove redundant edges (e.g. A –> (internal node) –> (internal node)) that arise from one case of the function mergeconstraintnodes!.

Redundant nodes in this case will always have only two edges while not being the root.

NetMerge.getsiblingcandidates — Method

A helper function that finds other nodes in the net that leaf belongs to that may be its sibling. This function sees past reticulations, so that e.g. in ((A,(B,#H1)),((C)#H1,D)) B and C will be marked as siblings.

NetMerge.inany — Method

Helper function that finds whether name is listed in any subset in subsets

NetMerge.internodedistance — Method

internodedistance(N::HybridNetwork; namelist::Union{Nothing,<:AbstractVector{String}}=nothing)

Calculates internode distances between all pairs of taxa in network N.

NetMerge.kconnected — Method

Takes minsubset, a subset defined in the function minimumuniquesubset, and, for each entry i of minsubset, gathers the k-1 taxa that i appears with most often across all quartets.

NetMerge.majorinternodedistance — Method

majorinternodedistance(N::HybridNetwork)

Calculates internode distances between all pairs of taxa in the major displayed tree of network N. If N is a tree, then internode distances are as expected for a tree.

NetMerge.mergesubnets! — Method

Merges two unrooted SubNets into a single unrooted SubNet Returns the merged subnet AND the edges in the merged net corresponding to where reticulations will be placed if there are corresponding reticulations discovered in the constraints.

IMPORTANT: this function should NOT make copies of edges or nodes, object references are stored and used in the main merging algorithm

NetMerge.minimumuniquesubset — Method

Finds the smallest subset S of names in namelist such that every quartet in R has at least one of its taxa in S.

NetMerge.netnj — Method

netnj(D::Matrix{Float64}, constraints::Vector{HybridNetwork})

Creates a super-network based on the constraint trees/networks in constraints and distance matrix D.

Arguments

D: distance matrix relating pairs of taxa. This can be generated from estimated gene trees with calculateAGID

NetMerge.parsequartets — Method

Reads output from MSCquartets in R and parses it so that it can be used efficiently with our julia code.

NetMerge.parsequartets — Method

Reads output from MSCquartets in R and parses it so that it can be used efficiently with our julia code.

NetMerge.placeretics! — Method

Updates net to include the reticulations that we've kept track of along the way in our algo but haven't placed yet.

NetMerge.removeredundantedges! — Method

Helper function that removes redundant edges in the graph that exist when some hybrids are pruned from the graph.

NetMerge.requiredhybridsubsets — Method

Current base subset decomposition schema. First, we find the smallest subset S of names such that all quartets have at least one taxa in the subset. Then, for each entry of S, we gather the set of 3 other taxa it appears with most often in the quartets in R.

The idea behind these subsets is to guarantee identifiability of hybridizations, hence why they are required.

NetMerge.updateconstraints! — Method

Updates constraint networks after (i, j) with names (nodenamei, nodenamej) have been merged. Also updates reticmap to keep track of any reticulations that get removed in this process.

By convention we keep nodenamei and replace node names with nodenamej

NetMerge.vecreplace! — Method

Helper function for manipulating graph adjacency lists.